Move pruning stuff into it's own module inside the crate
