+++ bin/cli/src/commands/prune.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use citrea_storage_ops::{Pruner, PruningConfig};
+use citrea_storage_ops::pruning::{Pruner, PruningConfig};
 use sov_db::ledger_db::{LedgerDB, SharedLedgerOps};
 use sov_db::native_db::NativeDB;
 use sov_db::rocks_db_config::RocksdbConfig;
