+++ crates/storage-ops/src/pruning/components/evm.rs
