+++ crates/storage-ops/src/pruning/components/ledger.rs
