+++ crates/storage-ops/src/pruning/components/mod.rs
