+++ crates/storage-ops/src/pruning/criteria.rs
